Moved db update and init code from test.common.server to src.common.server
