+++ nodejs/src/drive-explorer/api-server/update-db.js
@@ -2,4 +2,4 @@ import './src.common.server/utils/appSettings/appEnvAutoload.js';
 import { AppDbVersionUpdater } from './src/utils/noSql/nedb/dbUpdater.js';
 
 const appDbVersionUpdater = new AppDbVersionUpdater();
-appDbVersionUpdater.assureDbUpToDate({});
+appDbVersionUpdater.assureDbIsUpToDate({});
